fix: add remapping for nested dependency
foundry seems to have issues with nested dependencies. We import VaultRegistry which imports IERC4626 from its own dependencies. Foundry fails to compute the correct path. By adding it to the root project's remappings we can work around the issue. If it's fixed in the future we should update the repo